Checkpoint: Split disk manager struct & methods into separate file, add a simple unit test that mocks the GKE API
