+++ go.mod
@@ -7,6 +7,7 @@ require (
 	golang.org/x/oauth2 v0.0.0-20210201163806-010130855d6c
 	google.golang.org/api v0.38.0
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+	k8s.io/api v0.17.0
 	k8s.io/apimachinery v0.17.0
 	k8s.io/client-go v0.17.0
 )
